Add custom event handler sample
