+++ event-handler/custom-identity-event-handler/src/main/java/org/wso2/carbon/identity/customhandler/internal/UserRegistrationCustomEventHandlerComponent.java
@@ -5,25 +5,25 @@ import org.apache.commons.logging.LogFactory;
 import org.osgi.framework.BundleContext;
 import org.osgi.service.component.ComponentContext;
 import org.wso2.carbon.identity.event.handler.AbstractEventHandler;
-import org.wso2.carbon.identity.customhandler.handler.CustomUserSelfRegistrationHandler;
+import org.wso2.carbon.identity.customhandler.handler.UserRegistrationCustomEventHandler;
 import org.wso2.carbon.user.core.service.RealmService;
 
 /**
- * @scr.component name="org.wso2.carbon.identity.customhandler.internal.CustomUserSelfRegistrationHandlerComponent" immediate="true"
+ * @scr.component name="org.wso2.carbon.identity.customhandler.internal.UserRegistrationCustomEventHandlerComponent" immediate="true"
  * @scr.reference name="realm.service"
  * interface="org.wso2.carbon.user.core.service.RealmService" cardinality="1..1"
  * policy="dynamic" bind="setRealmService" unbind="unsetRealmService"
  */
-public class CustomUserSelfRegistrationHandlerComponent {
+public class UserRegistrationCustomEventHandlerComponent {
 
-    private static Log log = LogFactory.getLog(CustomUserSelfRegistrationHandlerComponent.class);
-    private CustomUserSelfRegistrationHandlerDataHolder dataHolder = CustomUserSelfRegistrationHandlerDataHolder
+    private static Log log = LogFactory.getLog(UserRegistrationCustomEventHandlerComponent.class);
+    private UserRegistrationCustomEventHandlerDataHolder dataHolder = UserRegistrationCustomEventHandlerDataHolder
             .getInstance();
 
     protected void activate(ComponentContext context) {
         try {
             BundleContext bundleContext = context.getBundleContext();
-            bundleContext.registerService(AbstractEventHandler.class.getName(), new CustomUserSelfRegistrationHandler(),
+            bundleContext.registerService(AbstractEventHandler.class.getName(), new UserRegistrationCustomEventHandler(),
                     null);
         } catch (Exception e) {
             log.error("Error while activating custom User selfRegistration handler component.", e);
